refactor(daytona): split lib.rs into client, state, tools modules
Split 1900-line monolith into focused modules:
- lib.rs (~185 lines): plugin registration, constants, capability
- client.rs (~490 lines): DaytonaClient HTTP client, urlencoding, client tests
- state.rs (~185 lines): types, state management helpers, serialization tests
- tools.rs (~870 lines): 7 tool implementations, tool tests

All 35 tests still pass. No behavioral changes.
