+++ integrations/daytona/src/client.rs
@@ -1,0 +1,644 @@
+//! Daytona API client and URL encoding utilities.
+
+use serde_json::{Value, json};
+use std::time::Duration;
+use tracing::debug;
+
+use crate::state::{ExecResult, SandboxInfo};
+use crate::{SANDBOX_READY_MAX_WAIT, SANDBOX_READY_POLL_INTERVAL};
+
+// ============================================================================
+// DaytonaClient - HTTP client for Daytona APIs
+// ============================================================================
+
+pub struct DaytonaClient {
+    http: reqwest::Client,
+    api_key: String,
+    api_base: String,
+    toolbox_base: String,
+}
+
+impl DaytonaClient {
+    pub fn new(api_key: String) -> Self {
+        Self::with_base_urls(
+            api_key,
+            crate::DAYTONA_API_BASE.to_string(),
+            crate::DAYTONA_TOOLBOX_BASE.to_string(),
+        )
+    }
+
+    pub fn with_base_urls(api_key: String, api_base: String, toolbox_base: String) -> Self {
+        Self {
+            http: reqwest::Client::new(),
+            api_key,
+            api_base,
+            toolbox_base,
+        }
+    }
+
+    // --- Generic request helpers ---
+
+    async fn management_request(
+        &self,
+        method: reqwest::Method,
+        path: &str,
+        body: Option<Value>,
+    ) -> Result<Value, String> {
+        let url = format!("{}{}", self.api_base, path);
+        let mut req = self
+            .http
+            .request(method, &url)
+            .bearer_auth(&self.api_key)
+            .header("Content-Type", "application/json");
+
+        if let Some(b) = body {
+            req = req.json(&b);
+        }
+
+        let resp = req
+            .send()
+            .await
+            .map_err(|e| format!("Failed to connect to Daytona API: {e}"))?;
+
+        let status = resp.status();
+        let body_text = resp
+            .text()
+            .await
+            .map_err(|e| format!("Failed to read response: {e}"))?;
+
+        if !status.is_success() {
+            return Err(format!("Daytona API error ({status}): {body_text}"));
+        }
+
+        if body_text.is_empty() {
+            return Ok(json!({}));
+        }
+
+        serde_json::from_str(&body_text).map_err(|e| format!("Invalid JSON from Daytona: {e}"))
+    }
+
+    /// Make a request to the Toolbox API (in-sandbox HTTP REST API).
+    async fn toolbox_request(
+        &self,
+        method: reqwest::Method,
+        sandbox_id: &str,
+        path: &str,
+        body: Option<Value>,
+    ) -> Result<Value, String> {
+        let url = format!("{}/{sandbox_id}{path}", self.toolbox_base);
+        let mut req = self
+            .http
+            .request(method, &url)
+            .bearer_auth(&self.api_key)
+            .header("Content-Type", "application/json");
+
+        if let Some(b) = body {
+            req = req.json(&b);
+        }
+
+        let resp = req
+            .send()
+            .await
+            .map_err(|e| format!("Failed to connect to sandbox: {e}"))?;
+
+        let status = resp.status();
+        let body_text = resp
+            .text()
+            .await
+            .map_err(|e| format!("Failed to read sandbox response: {e}"))?;
+
+        if !status.is_success() {
+            return Err(format!("Sandbox API error ({status}): {body_text}"));
+        }
+
+        if body_text.is_empty() {
+            return Ok(json!({}));
+        }
+
+        serde_json::from_str(&body_text).map_err(|e| format!("Invalid JSON from sandbox: {e}"))
+    }
+
+    /// Raw toolbox GET that returns bytes (for file download).
+    async fn toolbox_download(&self, sandbox_id: &str, path: &str) -> Result<Vec<u8>, String> {
+        let url = format!("{}/{sandbox_id}{path}", self.toolbox_base);
+        let resp = self
+            .http
+            .get(&url)
+            .bearer_auth(&self.api_key)
+            .send()
+            .await
+            .map_err(|e| format!("Failed to connect to sandbox: {e}"))?;
+
+        let status = resp.status();
+        if !status.is_success() {
+            let body_text = resp
+                .text()
+                .await
+                .map_err(|e| format!("Failed to read response: {e}"))?;
+            return Err(format!("Sandbox API error ({status}): {body_text}"));
+        }
+
+        resp.bytes()
+            .await
+            .map(|b| b.to_vec())
+            .map_err(|e| format!("Failed to read file bytes: {e}"))
+    }
+
+    /// Multipart file upload to the toolbox API.
+    async fn toolbox_upload(
+        &self,
+        sandbox_id: &str,
+        remote_path: &str,
+        content: &[u8],
+        filename: &str,
+    ) -> Result<(), String> {
+        let url = format!(
+            "{}/{sandbox_id}/files/upload?path={}",
+            self.toolbox_base,
+            urlencoding::encode(remote_path)
+        );
+
+        let part = reqwest::multipart::Part::bytes(content.to_vec())
+            .file_name(filename.to_string())
+            .mime_str("application/octet-stream")
+            .map_err(|e| format!("Failed to create multipart part: {e}"))?;
+
+        let form = reqwest::multipart::Form::new().part("file", part);
+
+        let resp = self
+            .http
+            .post(&url)
+            .bearer_auth(&self.api_key)
+            .multipart(form)
+            .send()
+            .await
+            .map_err(|e| format!("Failed to upload file: {e}"))?;
+
+        let status = resp.status();
+        if !status.is_success() {
+            let body_text = resp
+                .text()
+                .await
+                .map_err(|e| format!("Failed to read response: {e}"))?;
+            return Err(format!("File upload error ({status}): {body_text}"));
+        }
+
+        Ok(())
+    }
+
+    // --- Management API ---
+
+    pub async fn create_sandbox(&self, body: Value) -> Result<SandboxInfo, String> {
+        let resp = self
+            .management_request(reqwest::Method::POST, "/sandbox", Some(body))
+            .await?;
+        serde_json::from_value(resp).map_err(|e| format!("Failed to parse sandbox info: {e}"))
+    }
+
+    pub async fn get_sandbox(&self, sandbox_id: &str) -> Result<SandboxInfo, String> {
+        let resp = self
+            .management_request(
+                reqwest::Method::GET,
+                &format!("/sandbox/{sandbox_id}"),
+                None,
+            )
+            .await?;
+        serde_json::from_value(resp).map_err(|e| format!("Failed to parse sandbox info: {e}"))
+    }
+
+    pub async fn start_sandbox(&self, sandbox_id: &str) -> Result<(), String> {
+        self.management_request(
+            reqwest::Method::POST,
+            &format!("/sandbox/{sandbox_id}/start"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+
+    pub async fn stop_sandbox(&self, sandbox_id: &str) -> Result<(), String> {
+        self.management_request(
+            reqwest::Method::POST,
+            &format!("/sandbox/{sandbox_id}/stop"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+
+    pub async fn delete_sandbox(&self, sandbox_id: &str) -> Result<(), String> {
+        self.management_request(
+            reqwest::Method::DELETE,
+            &format!("/sandbox/{sandbox_id}"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+
+    pub async fn set_autostop(
+        &self,
+        sandbox_id: &str,
+        interval_minutes: u64,
+    ) -> Result<(), String> {
+        self.management_request(
+            reqwest::Method::POST,
+            &format!("/sandbox/{sandbox_id}/autostop/{interval_minutes}"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+
+    /// Wait for sandbox to reach "started" state after creation.
+    pub async fn wait_for_ready(&self, sandbox_id: &str) -> Result<(), String> {
+        let start = std::time::Instant::now();
+        let mut interval = SANDBOX_READY_POLL_INTERVAL;
+
+        while start.elapsed() < SANDBOX_READY_MAX_WAIT {
+            match self.get_sandbox(sandbox_id).await {
+                Ok(info) => {
+                    if info.state == "started" {
+                        debug!("Sandbox ready (state: started) after {:?}", start.elapsed());
+                        return Ok(());
+                    }
+                    if info.state == "error" || info.state == "build_failed" {
+                        return Err(format!("Sandbox entered error state: {}", info.state));
+                    }
+                    debug!("Sandbox not ready yet (state: {}), retrying...", info.state);
+                }
+                Err(e) => {
+                    debug!("Failed to poll sandbox state: {e}, retrying...");
+                }
+            }
+
+            tokio::time::sleep(interval).await;
+            interval = std::cmp::min(interval * 2, Duration::from_secs(4));
+        }
+
+        Err(format!(
+            "Sandbox did not become ready within {}s",
+            SANDBOX_READY_MAX_WAIT.as_secs()
+        ))
+    }
+
+    // --- Toolbox API: Process ---
+
+    pub async fn exec(
+        &self,
+        sandbox_id: &str,
+        command: &str,
+        cwd: Option<&str>,
+        timeout_ms: Option<u64>,
+    ) -> Result<ExecResult, String> {
+        let mut body = json!({ "command": command });
+        if let Some(c) = cwd {
+            body["cwd"] = json!(c);
+        }
+        if let Some(t) = timeout_ms {
+            body["timeout"] = json!(t);
+        }
+        let resp = self
+            .toolbox_request(
+                reqwest::Method::POST,
+                sandbox_id,
+                "/process/execute",
+                Some(body),
+            )
+            .await?;
+        serde_json::from_value(resp).map_err(|e| format!("Failed to parse exec result: {e}"))
+    }
+
+    // --- Toolbox API: Files ---
+
+    pub async fn file_download(&self, sandbox_id: &str, path: &str) -> Result<Vec<u8>, String> {
+        let encoded = urlencoding::encode(path);
+        self.toolbox_download(sandbox_id, &format!("/files/download?path={encoded}"))
+            .await
+    }
+
+    pub async fn file_upload(
+        &self,
+        sandbox_id: &str,
+        path: &str,
+        content: &[u8],
+    ) -> Result<(), String> {
+        let filename = path.rsplit('/').next().unwrap_or("file");
+        self.toolbox_upload(sandbox_id, path, content, filename)
+            .await
+    }
+
+    pub async fn file_list(&self, sandbox_id: &str, path: &str) -> Result<Vec<Value>, String> {
+        let encoded = urlencoding::encode(path);
+        let resp = self
+            .toolbox_request(
+                reqwest::Method::GET,
+                sandbox_id,
+                &format!("/files/?path={encoded}"),
+                None,
+            )
+            .await?;
+
+        // Response is an array of file entries
+        match resp.as_array() {
+            Some(arr) => Ok(arr.clone()),
+            None => Ok(vec![resp]),
+        }
+    }
+
+    pub async fn file_delete(&self, sandbox_id: &str, path: &str) -> Result<(), String> {
+        let encoded = urlencoding::encode(path);
+        self.toolbox_request(
+            reqwest::Method::DELETE,
+            sandbox_id,
+            &format!("/files?path={encoded}"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+
+    pub async fn create_folder(
+        &self,
+        sandbox_id: &str,
+        path: &str,
+        mode: &str,
+    ) -> Result<(), String> {
+        let encoded_path = urlencoding::encode(path);
+        let encoded_mode = urlencoding::encode(mode);
+        self.toolbox_request(
+            reqwest::Method::POST,
+            sandbox_id,
+            &format!("/files/folder?path={encoded_path}&mode={encoded_mode}"),
+            None,
+        )
+        .await?;
+        Ok(())
+    }
+}
+
+// ============================================================================
+// URL encoding helper (inline, avoids extra dependency)
+// ============================================================================
+
+pub(crate) mod urlencoding {
+    /// Percent-encode a string for use in query parameters.
+    pub fn encode(input: &str) -> String {
+        let mut result = String::with_capacity(input.len() * 2);
+        for byte in input.bytes() {
+            match byte {
+                b'A'..=b'Z' | b'a'..=b'z' | b'0'..=b'9' | b'-' | b'_' | b'.' | b'~' => {
+                    result.push(byte as char);
+                }
+                _ => {
+                    result.push('%');
+                    result.push_str(&format!("{byte:02X}"));
+                }
+            }
+        }
+        result
+    }
+}
+
+// ============================================================================
+// Tests
+// ============================================================================
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+    use wiremock::matchers::{method, path};
+    use wiremock::{Mock, MockServer, ResponseTemplate};
+
+    #[tokio::test]
+    async fn test_client_create_sandbox() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sandbox"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+                "id": "sb_test123",
+                "name": "Test Sandbox",
+                "state": "started"
+            })))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.create_sandbox(json!({"name": "Test"})).await;
+        assert!(result.is_ok());
+        let info = result.unwrap();
+        assert_eq!(info.id, "sb_test123");
+        assert_eq!(info.name, Some("Test Sandbox".to_string()));
+    }
+
+    #[tokio::test]
+    async fn test_client_get_sandbox() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("GET"))
+            .and(path("/sandbox/sb_test"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+                "id": "sb_test",
+                "name": "Test",
+                "state": "started"
+            })))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.get_sandbox("sb_test").await;
+        assert!(result.is_ok());
+        let info = result.unwrap();
+        assert_eq!(info.state, "started");
+    }
+
+    #[tokio::test]
+    async fn test_client_stop_sandbox() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sandbox/sb_test/stop"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!({})))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.stop_sandbox("sb_test").await;
+        assert!(result.is_ok());
+    }
+
+    #[tokio::test]
+    async fn test_client_delete_sandbox() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("DELETE"))
+            .and(path("/sandbox/sb_test"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!({})))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.delete_sandbox("sb_test").await;
+        assert!(result.is_ok());
+    }
+
+    #[tokio::test]
+    async fn test_client_exec() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sb_test/process/execute"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
+                "result": "hello world\n",
+                "exitCode": 0
+            })))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.exec("sb_test", "echo hello world", None, None).await;
+        assert!(result.is_ok());
+        let exec_result = result.unwrap();
+        assert_eq!(exec_result.exit_code, 0);
+        assert_eq!(exec_result.result, "hello world\n");
+    }
+
+    #[tokio::test]
+    async fn test_client_file_list() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("GET"))
+            .and(path("/sb_test/files/"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(json!([
+                {"name": "main.py", "isDir": false, "size": 42},
+                {"name": "src", "isDir": true, "size": 0}
+            ])))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.file_list("sb_test", "/home/daytona").await;
+        assert!(result.is_ok());
+        let entries = result.unwrap();
+        assert_eq!(entries.len(), 2);
+    }
+
+    #[tokio::test]
+    async fn test_client_404_error() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("GET"))
+            .and(path("/sandbox/nonexistent"))
+            .respond_with(
+                ResponseTemplate::new(404).set_body_string("{\"error\":\"Sandbox not found\"}"),
+            )
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.get_sandbox("nonexistent").await;
+        assert!(result.is_err());
+        assert!(result.unwrap_err().contains("404"));
+    }
+
+    #[tokio::test]
+    async fn test_client_401_error() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sandbox"))
+            .respond_with(
+                ResponseTemplate::new(401).set_body_string("{\"error\":\"Unauthorized\"}"),
+            )
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "bad_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.create_sandbox(json!({"name": "Test"})).await;
+        assert!(result.is_err());
+        assert!(result.unwrap_err().contains("401"));
+    }
+
+    #[tokio::test]
+    async fn test_client_500_error() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sandbox/sb_test/start"))
+            .respond_with(ResponseTemplate::new(500).set_body_string("Internal Server Error"))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.start_sandbox("sb_test").await;
+        assert!(result.is_err());
+        assert!(result.unwrap_err().contains("500"));
+    }
+
+    #[tokio::test]
+    async fn test_client_malformed_response() {
+        let mock_server = MockServer::start().await;
+        Mock::given(method("POST"))
+            .and(path("/sandbox"))
+            .respond_with(ResponseTemplate::new(200).set_body_string("not valid json"))
+            .mount(&mock_server)
+            .await;
+
+        let client = DaytonaClient::with_base_urls(
+            "test_key".to_string(),
+            mock_server.uri(),
+            mock_server.uri(),
+        );
+        let result = client.create_sandbox(json!({"name": "Test"})).await;
+        assert!(result.is_err());
+        assert!(result.unwrap_err().contains("Invalid JSON"));
+    }
+
+    #[test]
+    fn test_encode_simple() {
+        assert_eq!(urlencoding::encode("hello"), "hello");
+    }
+
+    #[test]
+    fn test_encode_path_with_slashes() {
+        assert_eq!(
+            urlencoding::encode("/home/daytona/main.py"),
+            "%2Fhome%2Fdaytona%2Fmain.py"
+        );
+    }
+
+    #[test]
+    fn test_encode_path_with_spaces() {
+        let encoded = urlencoding::encode("/my project/file name.txt");
+        assert!(encoded.contains("%20"));
+    }
+}
